re-enable rest of package, minor rearrangements
