+++ src/SpectralKit.jl
@@ -7,11 +7,11 @@ using StaticArrays: MVector, SVector, sacollect
 
 include("utilities.jl")
 include("derivatives.jl")
-# include("domains.jl")
-# include("transformations.jl")
-# include("generic_api.jl")
-# include("chebyshev.jl")
-# include("smolyak_traversal.jl")
-# include("smolyak_api.jl")
+include("domains.jl")
+include("transformations.jl")
+include("generic_api.jl")
+include("chebyshev.jl")
+include("smolyak_traversal.jl")
+include("smolyak_api.jl")
 
 end # module
