+++ src/transformations.jl
@@ -2,9 +2,8 @@
 ##### transformations
 #####
 
-export transform_to, transform_from, coordinate_transformations,
+export domain, domain_kind, transform_to, transform_from, coordinate_transformations,
     BoundedLinear, InfRational, SemiInfRational
-
 
 ####
 #### generic api
@@ -118,14 +117,9 @@ function transform_to(domain::CoordinateDomains{T}, ct::CoordinateTransformation
     SVector(transform_to(domain, ct, _ntuple_like(T, x)))
 end
 
-function transform_to(domain::CoordinateDomains, ct::CoordinateTransformations, ∂x::∂Input)
-    transform_to(domain, ct, _lift(∂x))
-end
-
 function transform_to(domain::CoordinateDomains, ct::CoordinateTransformations,
-                      ∂x::∂InputLifted)
-    (; ∂specification, lifted_x) = ∂x
-    ∂InputLifted(∂specification, transform_to(domain, ct, lifted_x))
+                      Dx::∂CoordinateExpansion)
+    ∂CoordinateExpansion(x.∂D, transform_to(domain, ct, x.x))
 end
 
 function transform_from(domain::CoordinateDomains, ct::CoordinateTransformations, x::Tuple)
@@ -187,13 +181,13 @@ function transform_to(::PM1, t::BoundedLinear, y::Real)
     (y - m) / s
 end
 
-function transform_to(domain::PM1, t::BoundedLinear, y::Derivatives{N}) where N
+function transform_to(domain::PM1, t::BoundedLinear, y::𝑑Expansion{Dp1}) where Dp1
     (; m, s) = t
     (; derivatives) = y
     y0, yD... = derivatives
     x0 = transform_to(domain, t, y0)
     xD = map(y -> y / s, yD)
-    Derivatives((x0, xD...))
+    𝑑Expansion((x0, xD...))
 end
 
 function domain(t::BoundedLinear)
@@ -256,16 +250,16 @@ function transform_to(::PM1, t::SemiInfRational, y::Real)
     end
 end
 
-function transform_to(domain::PM1, t::SemiInfRational, y::Derivatives{N}) where N
+function transform_to(domain::PM1, t::SemiInfRational, y::𝑑Expansion{Dp1}) where Dp1
     (; A, L) = t
     (; derivatives) = y
     x0 = transform_to(domain, t, derivatives[1])
-    N == 1 && return Derivatives((x0, ))
+    Dp1 == 1 && return 𝑑Expansion(SVector(x0))
     # based on Boyd (2001), Table E.7
     Q = abs2(x0 - 1)
     x1 = (derivatives[2] * Q) / (2*L)
-    N == 2 && return Derivatives((x0, x1))
-    error("$(N-1)th derivative not implemented yet, open an issue.")
+    Dp1 == 2 && return 𝑑Expansion(SVector(x0, x1))
+    error("$(Dp1-1)th derivative not implemented yet, open an issue.")
 end
 
 function domain(t::SemiInfRational)
@@ -321,16 +315,16 @@ function transform_to(::PM1, t::InfRational, y::Real)
     end
 end
 
-function transform_to(domain::PM1, t::InfRational, y::Derivatives{N}) where N
+function transform_to(domain::PM1, t::InfRational, y::𝑑Expansion{Dp1}) where Dp1
     (; A, L) = t
     (; derivatives) = y
     x0 = transform_to(domain, t, derivatives[1])
-    N == 1 && return Derivatives((x0, ))
+    N == 1 && return Derivatives(SVector(x0))
     # based on Boyd (2001), Table E.5
     Q = 1 - abs2(x0)
     sQ = √Q
     x1 = (derivatives[2] * Q * sQ) / L
-    N == 2 && return Derivatives((x0, x1))
+    N == 2 && return 𝑑Expansion(SVector(x0, x1))
     error("$(N-1)th derivative not implemented yet, open an issue.")
 end
